+++ patch_api/api/projects_api.py
@@ -28,7 +28,19 @@ class ProjectsApi(object):
     Do not edit the class manually.
     """
 
-    ALLOWED_QUERY_PARAMS = ["mass_g", "price_cents_usd", "project_id", "page"]
+    ALLOWED_QUERY_PARAMS = [
+        "mass_g",
+        "price_cents_usd",
+        "project_id",
+        "page",
+        "distance_m",
+        "transportation_method",
+        "package_mass_g",
+        "create_order",
+        "model",
+        "make",
+        "year",
+    ]
 
     def __init__(self, api_client=None):
         self.api_client = api_client
@@ -94,6 +106,13 @@ class ProjectsApi(object):
         all_params.append("price_cents_usd")
         all_params.append("project_id")
         all_params.append("metadata")
+        all_params.append("distance_m")
+        all_params.append("transportation_method")
+        all_params.append("package_mass_g")
+        all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params:
@@ -214,6 +233,13 @@ class ProjectsApi(object):
         all_params.append("price_cents_usd")
         all_params.append("project_id")
         all_params.append("metadata")
+        all_params.append("distance_m")
+        all_params.append("transportation_method")
+        all_params.append("package_mass_g")
+        all_params.append("create_order")
+        all_params.append("make")
+        all_params.append("model")
+        all_params.append("year")
 
         for key, val in six.iteritems(local_var_params["kwargs"]):
             if key not in all_params: