+++ test/test_estimates_api.py
@@ -43,6 +43,61 @@ class TestEstimatesApi(unittest.TestCase):
         retrieved_estimate = self.api.retrieve_estimate(id=estimate.data.id)
         self.assertTrue(retrieved_estimate)
 
+    def test_create_and_retrieve_flight_estimate(self):
+        """Test case for create_flight_estimate
+
+        Create an estimate based on the distance in meters flown by an airplane # noqa: E501
+        """
+        distance_m = 10000000
+        estimate = self.api.create_flight_estimate(
+            distance_m=distance_m, create_order=True
+        )
+        self.assertEqual(estimate.data.type, "flight")
+        self.assertEqual(estimate.data.order.mass_g, 1032000)
+        self.assertEqual(estimate.data.mass_g, 1032000)
+
+        retrieved_estimate = self.api.retrieve_estimate(id=estimate.data.id)
+        self.assertTrue(retrieved_estimate)
+
+    def test_create_and_retrieve_shipping_estimate(self):
+        """Test case for create_shipping_estimate
+
+        Create an estimate based on the shipping distance, transportation method, and package mass  # noqa: E501
+        """
+        distance_m = 10000000
+        package_mass_g = 1000
+        transportation_method = "sea"
+        estimate = self.api.create_shipping_estimate(
+            distance_m=distance_m,
+            package_mass_g=package_mass_g,
+            transportation_method=transportation_method,
+            create_order=False,
+        )
+        self.assertEqual(estimate.data.order, None)
+        self.assertEqual(estimate.data.type, "shipping")
+        self.assertEqual(estimate.data.mass_g, 373)
+
+        retrieved_estimate = self.api.retrieve_estimate(id=estimate.data.id)
+        self.assertTrue(retrieved_estimate)
+
+    def test_create_and_retrieve_vehicle_estimate(self):
+        """Test case for create_vehicle_estimate
+
+        Create an estimate based on the vehicle distance, transportation method, and package mass  # noqa: E501
+        """
+        distance_m = 10000000
+        make = "Toyota"
+        model = "Corolla"
+        year = 1995
+        estimate = self.api.create_vehicle_estimate(
+            distance_m=distance_m, model=model, make=make, year=year
+        )
+        self.assertEqual(estimate.data.type, "vehicle")
+        self.assertEqual(estimate.data.mass_g, 5719674)
+
+        retrieved_estimate = self.api.retrieve_estimate(id=estimate.data.id)
+        self.assertTrue(retrieved_estimate)
+
 
 if __name__ == "__main__":
     unittest.main()