+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-02-08
+
+### Added
+
+- Adds support for creating Shipping/Flight/Vehicle estimates.
+- Sets the library version to 1.3.0 to track the [Patch Ruby SDK](https://github.com/patch-technology/patch-ruby) and [Patch Node SDK](https://github.com/patch-technology/patch-node).
+
 ## [1.0.1] - 2021-01-22
 
 ### Added
 
 - Pre-release of v1 Library
 - Adds support for Orders, Estimates, Projects and Preferences
-